+++ CHANGELOG.md
@@ -24,6 +24,12 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - MerchantInstallment::Log sub-resource
 - MerchantPurchase resource
 - MerchantPurchase::Log sub-resource
+- SplitReceiver resource
+- SplitReceiver::Log sub-resource
+- SplitProfile resource
+- SplitProfile::Log sub-resource
+- Split resource
+- Split::Log sub-resource
 
 ### Fixed
 - Docstrings across Boleto, BoletoPayment, BoletoHolmes, BrcodePayment, DarfPayment, Deposit, DictKey,